+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 ## [1.0.1] - 2024-11-13
 
  - Added Invoice endpoint
+ - Added Ordering endpoint